+++ CHANGELOG.md
@@ -13,6 +13,10 @@ Read: ~100 records/ms
 - Add a "revisions" mode to history.  Now history records can be set per row instead of database wide.
 - Add `revisionComment` function to standard queries allowing you to add information about each revision into the datastore.
 
+## [0.8.6] 5-25-2017
+- The NanoSQLInstance method `random16bits` is now public instead of private.
+- Fixed an issue with the ORM not correctly removing entries.
+
 ## [0.8.54] 5-23-2017
 - Fixed an issue where CSV export would skip undefined/false columns entirely, breaking the column alignment.